rename package from cloudwatch to cwlogs
