+++ internal/aws/cwlogs/cwlog_client.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cloudwatch // import "github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cloudwatch"
+package cwlogs // import "github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cwlogs"
 
 import (
 	"fmt"
@@ -27,7 +27,7 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cloudwatch/handler"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cwlogs/handler"
 )
 
 var collectorDistribution = "opentelemetry-collector-contrib"
