+++ internal/aws/cwlogs/go.mod
@@ -1,4 +1,4 @@
-module github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cloudwatch
+module github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cwlogs
 
 go 1.17
