+++ internal/aws/cwlogs/handler/request_structuredlog_handler.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package handler // import "github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cloudwatch/handler"
+package handler // import "github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cwlogs/handler"
 
 import "github.com/aws/aws-sdk-go/aws/request"
