+++ internal/aws/cwlogs/handler/request_structuredlog_handler_test.go
